Fix health not being reapplied when the player respawns.
When players respawn they will now attempt to respawn with their former
maximum health. For example if keep inventory is set to true and you had
40 health before you died, you will now respawn with 40 health. This issue
was reported in #7.

This also unintentionally fixes the issue reported in #5 as teleporting
from the end to the overworld respawns the player behind the scenes. They
do this to allow the player to read the end game poem and credits without
being killed or messed with by other players.